+++ src/user/user.service.ts
@@ -21,4 +21,11 @@ export class UserService {
   async update(id: string, data: Prisma.UserUpdateInput): Promise<User> {
     return this.prisma.user.update({ where: { id }, data });
   }
+
+  async updateByEmail(
+    email: string,
+    data: Prisma.UserUpdateInput,
+  ): Promise<User> {
+    return this.prisma.user.update({ where: { email }, data });
+  }
 }